Disallow deletion and update of tutor(s) if tutor is assigned to a department already.

+++ src/main/java/com/swiftpot/timetable/controllers/TutorController.java
@@ -6,6 +6,7 @@ import com.swiftpot.timetable.model.SuccessfulOutgoingPayload;
 import com.swiftpot.timetable.repository.DepartmentDocRepository;
 import com.swiftpot.timetable.repository.TutorDocRepository;
 import com.swiftpot.timetable.repository.db.model.TutorDoc;
+import com.swiftpot.timetable.services.GeneralAbstractServices;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
@@ -26,6 +27,8 @@ public class TutorController {
     TutorDocRepository tutorDocRepository;
     @Autowired
     DepartmentDocRepository departmentDocRepository;
+    @Autowired
+    GeneralAbstractServices generalAbstractServices;
 
     @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
     private OutgoingPayload createTutor(@RequestBody TutorDoc tutorDoc) {
@@ -57,24 +60,33 @@ public class TutorController {
         }
     }
 
-    @RequestMapping(path = "/{id}",method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
+    @RequestMapping(path = "/{id}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
     private OutgoingPayload updateTutor(@PathVariable String id,
                                         @RequestBody TutorDoc tutorDoc) {
         if (tutorDocRepository.exists(id)) {
-            tutorDoc.setId(id);
-            TutorDoc tutorDocUpdated = tutorDocRepository.save(tutorDoc);
-            return new SuccessfulOutgoingPayload(tutorDocUpdated);
+            if (generalAbstractServices.isTutorAssignedToAnyDepartment(tutorDoc) == true) {
+                return new ErrorOutgoingPayload("Tutor Can not be updated because tutor is already assigned to a department.\n To update this tutor,remove tutor from the assigned department");
+            } else {
+                tutorDoc.setId(id);
+                TutorDoc tutorDocUpdated = tutorDocRepository.save(tutorDoc);
+                return new SuccessfulOutgoingPayload(tutorDocUpdated);
+            }
         } else {
             return new ErrorOutgoingPayload("Id does not exist");
         }
     }
 
-    @RequestMapping(path="/{id}",method = RequestMethod.DELETE)
+    @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
     private OutgoingPayload deleteTutor(@PathVariable String id) {
         if (tutorDocRepository.exists(id)) {
-            tutorDocRepository.delete(id);
-            return new SuccessfulOutgoingPayload("Deleted Successfully");
-        }else {
+            TutorDoc tutorDocToDelete = tutorDocRepository.findOne(id);
+            if (generalAbstractServices.isTutorAssignedToAnyDepartment(tutorDocToDelete) == true) {
+                return new ErrorOutgoingPayload("Tutor Can not be deleted because tutor is already assigned to a department.\n To delete this tutor,remove tutor from the assigned department");
+            } else {
+                tutorDocRepository.delete(id);
+                return new SuccessfulOutgoingPayload("Deleted Successfully");
+            }
+        } else {
             return new ErrorOutgoingPayload("Id does not exist");
         }
     }
@@ -82,10 +94,30 @@ public class TutorController {
     @RequestMapping(method = RequestMethod.DELETE)
     private OutgoingPayload deleteAllTutors() {
         if (tutorDocRepository.count() > 1) {
-            tutorDocRepository.deleteAll();
-            return new SuccessfulOutgoingPayload("Deleted Successfully");
-        }else {
+            List<TutorDoc> allTutorDocsInDb = tutorDocRepository.findAll();
+            if (isAnyTutorAssignedToADepartment(allTutorDocsInDb)) {
+                return new ErrorOutgoingPayload("All tutors cannot be deleted because some tutors are assigned to departments.\n Remove all tutors from departments if you want all tutors to be deleted");
+            } else {
+                tutorDocRepository.deleteAll();
+                return new SuccessfulOutgoingPayload("Deleted Successfully");
+            }
+        } else {
             return new ErrorOutgoingPayload("No Tutors To Delete Currently");
         }
+    }
+
+    /**
+     * @param tutorDocs
+     * @return
+     */
+    protected boolean isAnyTutorAssignedToADepartment(List<TutorDoc> tutorDocs) {
+        boolean isAnyTutorAssignedToADepartment = false;
+        for (int i = 0; i < tutorDocs.size(); i++) {
+            if (generalAbstractServices.isTutorAssignedToAnyDepartment(tutorDocs.get(i))) {
+                isAnyTutorAssignedToADepartment = true;
+                break;
+            }
+        }
+        return isAnyTutorAssignedToADepartment;
     }
 }
